updated/corrected physics in cooling, eos routine
corrected the way that number density was calculated (which was based on
an ISM type gas) in sph.c.

also changed the root-find interval in call to newtraph. newtraph has no
fail mechanism other than Msgf's (which I don't know where those are
printed to, it's not anywhere (like outfiles) that I can see). so it is
completely possible, that this rootfind has never worked with my setups
and I did not know about it.