fiddling with btab vs p issue in burning
use p, not btab. This works now.

btab is a pointer to an array of SPHbody structures; i.e. btab always
points to the first array element. Thus must use p to access each
particle. this actually works with the abund member if one actually
iterates only to element NISO, not Nel (which is not equal to NISO
anymore, duh).
but somewhere, there is still a segmentation violation (solven??)