changed abundances to x in solven
pass in mass fractions (as xin), they are converted to nucleon fractions
by solven. pass out mass fractions.
unfortunately, in this mode of the code the 'global' arrays are only seen
in fortran, not C. and also, cannot pass in an array that is in a common
block elsewhere. hence need an extra array, xin, that is used for passing
the abundances in and out of the subroutine. so at the end of the code,
a necessary step is setting xin = x again.