another NISO vs NNETW issue, plus updated some makefiles
NISO= number is isotopes tracked, currently 22
NNETW= number of isotopes in the network, currently 20
the initial sdf files used to be populated with 22 isotopes, but I
recently changed that to 20. Now the zero's in the extra columns seem
to be causing problems. Need to fix this issue some day....

hopefully made it easier to keep makefiles up to date when compilers
on clusters are updated