improving the cooling routine, yet again
I decided to try to make the switch for cooling dependend on a
so-called photospheric radius, calculated as R0-mfp_i, where R0
is the size of the simulation, and mfp_i is the mean free path
calculated for particle i. This could become problematic again
when a medium is added around the star. Also possibly since mfp
depends on particle. But hopefully this is slightly better than
comparing the mfp to the particle scale length.