adjusted some compiler flags in Make.tacc and updated some things
this arose out of an attempt at getting a working executable compiled
with the gcc compilers (with Make.tacc-gcc, which creates an executable,
but which then crashes when building the network libraries.). In the
course of this attempt some function definitions were updated, as
well as some 'if' checks updated. Make.tacc (using intel+openmpi)
creates a running executable (which appears to be running a bit slower
than on saguaro, but maybe that's just me) for SNSPH. However, WVT
needs to be built with .tacc-gcc (using gcc+openmpi(intel)), since the
sequential compile with pure intel generates a __intel_cpu_indicator_init
related link-error (for newer versions than openmpi-1.2.7).

Another change was including the electron number density in eos_n for
calculating u and temp (again?). I think I've concluded that this more
accurately reproduces the temperature in the original progenitor
profile (although it really should not make a difference).

(also, I need to stop experimenting in the master-branch)