attempting to improve efficiency of the code
1) convert physical constants to user-units once in main only,
instead of every time they're needed. they are global variables
that can be used by #including units.h
2) calculate often used exponentials (e.g. inverses, squares, ...)
of the conversion factors once in main, and treat those as global
variables (in units.h) as well. the idea is to avoid divisions
when iterating over the particles.

don't know if it will make a difference.... but it compiles

also, fixed cooling to only turn on at 0.5 yrs after the explosion.
radiative losses in SNRs only become important once they enter
the Sedov-Taylor phase. so the cooling routine might not have a big
impact on the evolution in any case? but at least it could be used
to get the ionization state of the material, assuming whatever
conditions the cooling table and ionization fractions are valid for.