fixing an error in the calculation of pressure and sound speed
pgas was multiplied erroneously by 1.5

In his email with how to calculate pressure, Chris said to use the material
eos to calculate the sound speed. for some reason I used the total pressure
to calculate the sound speed.

perhaps this is a reason why the runs since then take forever??