change A_COEFF to A_RAD, calculate temp in cgs units
to avoid roundoff to zero errors. in code units, physical constants
can take on (typically) VERY small values, which makes things zero
that shouldn't be. as a result, gas pressure has been zero in the
calculations so far (this is ok for a star while it explodes, but
not afterwards). Also, this resulted in a not defined sound speed
(after that calculation was changed from using p->pr to pgas).