new condition for accreting particles onto the central mass
trying to impose a minimum accretion radius by calculating a standard
NS radius of 10km and a "Schwartzschild radius" for speeds of 0.1c.
Then the bndry.r is set to the biggest one. The growing algorithm
remains the same, so the bndry can increase beyond that if necessary.
The new condition also accretes particles that are moving in too fast
(currently 2Rsol/100sec) to prevent crashes caused by fast particles
that weren't eaten but probably should have been.