adding external pressure term, and switch for using rho_est in prep_cool_burn
The external pressure term is a constant value set in the ctl file (where
it can also be switched off). It is subtracted from all particles
outside ~97.5% of sysradius (R0). So far performs reasonably well.

Since prep_cool_burn is called in both update_intermediate and
update_final, but rho_est is used in the former and rho in the latter,
I added a switch to select which rho is used (0= rho, 1= rho_est).
For consistency.