AdjustBtab4: implementing partial particle accretion, as in TsingWai's code
Once particles are within a smoothing length of the bndry -radius, an amount
of mass proportional to the free-fall timescale (dt/tff) is accreted. The
purpose of this is to mimick a smoother accretion. One change I made is to
continue this partial accretion as the SPH particle moves inside of the bndry -
radius (up to one smoothing length past it). Another change I made is to
check for SPH particles close to the bndry that are falling in at high speeds
(currently 0.033*c_light) and eat those as well (whole), to prevent SPH particles
from being able to 'jump over' the bndry and effectively get ejected at high
speeds (this is likely unphysical, but a more immediate concern was that this
crashes the run), which this seems to accomplish usually.