AdjustBtab4: adjustment to the particle eating algorithm
do not accrete particles based on their infalling velocities. occasionally
particles seem to be able to 'jump' over the absorbing boundary and effectively
get ejected at high speeds (several ord. of mag. over the typical velocity);
this, of course, crashes the run. if the run crashes, increase the bndry_r
(e.g. double it) and restart the run. in all tested runs this has performed
reasonably well (i.e. at most one restart was required).

the factor by which the bndry_r is increased was changed to 0.9 (from 0.5 or
so)in order to accrete faster. this possibly gets closer to the CCO masses
calculated in Chris' 1D code.