some adjustment to the main makefile used on Lonestar
the svml library is linked by default in the intel compilers vers. 11.1 and
greater, linking it again links it dynamically which results in the
'intel_cpu_indicator_init' error (possibly a mixing of compilers might have
contributed to that too).

also turning on parallel IO, which seems to work on Lonestar.