implement read/write functionality of different "types" of SDF files
dependent upon which physics packages are turned on. Thus,
cooling/burning, which require knowledge of the chemical composition,
also turns on reading/writing of SDF files with abundance information
(currently hardcoded to 20 isotopes). The "basic" mode for SN
calculations reads/writes SDF files with no abundance information
(though it can probably read a file with abundance information, but
won't propagate that info to the output).